requirements.txt
dlib
cmake
numpy
paho-mqtt